PE #18 + #67
This is sort of a dynamic programming solution-- though not formally.
It's just an in-order traversal, but each node holds the large accumulated
value yet vistied upon it. If the route currently under consideration
doesn't exceed the previous max (if exists) then it'll get pruned.
Runs in 200ms on the big one.