PE20 - without a bignum
I've attempted this problem a bunch of times. It's easy with a bignum
but I've always want to avoid using it.

The solution leverages this property:
ab mod n = [(a mod n)(b mod n)] mod n

I didn't come up with a good way to measure how far to iterate. A bunch
of "leading zeros" heuristically tells you when to stop. I just googled
how many digits there are in 100!